# edge/medbook-bff/src/main/resources/openapi/bff-api_v1.0.yaml
openapi: 3.0.3
...
info:
  title: MedBook - BFF API
  description: |
    ## MedBook BFF
    Unico punto di ingresso del frontend verso il sistema.
    Orchestra le chiamate ai DMN via Feign client e restituisce risposte aggregate.
    Non ha database proprio. E' un orchestratore stateless.

    ### Ruoli supportati
    - ROLE_PATIENT: paziente registrato
    - ROLE_DOCTOR: medico
    - ROLE_RECEPTIONIST: addetto alla reception
    - ROLE_ADMIN: amministratore

  version: 1.0.0
  contact:
    name: MedBook Team
    email: djizalfra@gmail.com
servers:
  - url: http://localhost:8090
    description: BFF locale - sviluppo
...
paths:
...
  /bff/v1/clinics:

    post:
      tags:
        - Clinics
      operationId: postCreateClinic
      summary: Crea una nuova sede clinica
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClinicBffRequest'
      responses:
        '201':
          description: Sede creata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiResponse'

    get:
      tags:
        - Clinics
      operationId: getAllClinics
      summary: Lista sedi cliniche
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: city
          in: query
          required: false
          schema:
            type: string
        - name: name
          in: query
          required: false
          schema:
            type: string
        - name: email
          in: query
          required: false
          schema:
            type: string
        - name: province
          in: query
          required: false
          schema:
            type: string
        - name: phone
          in: query
          required: false
          schema:
            type: string
        - name: address
          in: query
          required: false
          schema:
            type: string
        - name: postalCode
          in: query
          required: false
          schema:
            type: string
        - name: createdFrom
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: createdTo
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: updatedFrom
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: updatedTo
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Lista sedi recuperata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiResponse'
  /bff/v1/clinics/{clinicId}:

    get:
      tags:
        - Clinics
      operationId: getClinicById
      summary: Dettaglio sede clinica
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
        - $ref: '#/components/parameters/ClinicIdPathParam'
      responses:
        '200':
          description: Sede trovata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiResponse'
        '404':
          description: Sede non trovata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiErrorResponse'

    patch:
      tags:
        - Clinics
      operationId: patchUpdateClinic
      summary: Aggiorna dati sede clinica
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
        - $ref: '#/components/parameters/ClinicIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateClinicBffRequest'
      responses:
        '200':
          description: Sede aggiornata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiVoidResponse'

    delete:
      tags:
        - Clinics
      operationId: deleteClinic
      summary: Elimina logicamente una sede clinica
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
        - $ref: '#/components/parameters/ClinicIdPathParam'
      responses:
        '200':
          description: Sede eliminata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiVoidResponse'
  /bff/v1/clinics/{clinicId}/restore:

    patch:
      tags:
        - Clinics
      operationId: patchRestoreClinic
      summary: Ripristina una sede clinica eliminata logicamente
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
        - $ref: '#/components/parameters/ClinicIdPathParam'
      responses:
        '200':
          description: Sede ripristinata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiVoidResponse'
        '404':
          description: Sede non trovata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiErrorResponse'
        '409':
          description: La sede non e cancellata - impossibile ripristinare
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiErrorResponse'

  # ===========================================================================
  # ASSIGNMENTS
  # ===========================================================================
  /bff/v1/clinics/{clinicId}/assignments:

    post:
      tags:
        - Assignments
      operationId: postCreateAssignment
      summary: Assegna un medico a una sede
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
        - $ref: '#/components/parameters/ClinicIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssignmentBffRequest'
      responses:
        '201':
          description: Assegnazione creata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiResponse'

    get:
      tags:
        - Assignments
      operationId: getAllAssignments
      summary: Lista assegnazioni medici di una sede
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
        - $ref: '#/components/parameters/ClinicIdPathParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
        - name: doctorId
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Lista assegnazioni recuperata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiResponse'
  /bff/v1/clinics/{clinicId}/assignments/{assignmentId}:

    patch:
      tags:
        - Assignments
      operationId: patchUpdateAssignment
      summary: Aggiorna un'assegnazione
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
        - $ref: '#/components/parameters/ClinicIdPathParam'
        - name: assignmentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAssignmentBffRequest'
      responses:
        '200':
          description: Assegnazione aggiornata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiVoidResponse'

    delete:
      tags:
        - Assignments
      operationId: deleteAssignment
      summary: Rimuove un'assegnazione
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
        - $ref: '#/components/parameters/ClinicIdPathParam'
        - name: assignmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Assegnazione rimossa con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiVoidResponse'
  /bff/v1/clinics/{clinicId}/assignments/{assignmentId}/restore:

    patch:
      tags:
        - Assignments
      operationId: patchRestoreAssignment
      summary: Ripristina un'assegnazione eliminata logicamente
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
        - $ref: '#/components/parameters/ClinicIdPathParam'
        - name: assignmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Assegnazione ripristinata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiVoidResponse'
        '404':
          description: Sede o assegnazione non trovati
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiErrorResponse'
        '409':
          description: L'assegnazione non e cancellata - impossibile ripristinare
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiErrorResponse'

  # ===========================================================================
  # AVAILABILITY SEARCH (logica BFF)
  # ===========================================================================
...
components:
...
  parameters:
    MedBookContextParam:
      name: X-MedBook-Context
      in: header
      required: false
      description: |
        Contesto applicativo della richiesta - iniettato da MedBookContextInterceptor.
        Non viene letto dall'header HTTP: il resolver preleva MedBookContextHolder.get().
      schema:
        $ref: '#/components/schemas/MedBookContext'
...
    ClinicIdPathParam:
      name: clinicId
      in: path
      required: true
      schema:
        type: string
        example: CLN-1
...
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
    SizeParam:
      name: size
      in: query
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    SortParam:
      name: sort
      in: query
      required: false
      schema:
        type: string
        example: createdAt,desc

  # ---------------------------------------------------------------------------
  # SCHEMAS
  # ---------------------------------------------------------------------------
  schemas:
...
    MedBookApiResponse:
      type: object
      description: Risposta di successo con payload.
      properties:
        httpStatus:
          type: integer
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
        data:
          description: Payload della risposta
        page:
          $ref: '#/components/schemas/MedBookPageResponse'
    MedBookApiVoidResponse:
      type: object
      description: Risposta base per endpoint senza payload.
      properties:
        httpStatus:
          type: integer
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
    MedBookApiErrorResponse:
      type: object
      description: Formato standard delle risposte di errore.
      properties:
        httpStatus:
          type: integer
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
        errorCode:
          type: string
        message:
          description: Messaggio di errore
    MedBookPageResponse:
      type: object
      description: Metadati di paginazione.
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    MedBookContext:
      type: object
      description: Contesto della richiesta HTTP.
      properties:
        traceId:
          type: string
        username:
          type: string
        roles:
          type: array
          items:
            type: string
        requestTimestamp:
          type: string
          format: date-time

    # =========================================================================
    # PATIENT REQUEST
    # =========================================================================
...
    CreateClinicBffRequest:
      type: object
      description: Dati per la creazione di una nuova sede clinica
      required:
        - name
        - email
        - phone
        - address
        - city
        - postalCode
        - province
        - termsAccepted
      properties:
        name:
          type: string
          example: MedBook Milano Centro
        email:
          type: string
          format: email
          example: milano.centro@medbook.it
        phone:
          type: string
          example: "+390212345678"
        address:
          type: string
          example: Via Roma 1
        city:
          type: string
          example: Milano
        postalCode:
          type: string
          example: "20121"
        province:
          type: string
          maxLength: 100
          description: Sigla provincia (2 caratteri, es. TO, MI, NA)
          example: TO
        termsAccepted:
          type: boolean
          description: |
            Conferma accettazione dei Termini di Servizio da parte della clinica.
            Deve essere true per procedere con la registrazione.
          example: true
    UpdateClinicBffRequest:
      type: object
      description: Dati per l'aggiornamento parziale di una sede clinica
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        postalCode:
          type: string
        province:
          type: string
          maxLength: 100
          description: Sigla provincia (2 caratteri, es. TO, MI, NA)
          example: TO
        status:
          type: string

    # =========================================================================
    # ASSIGNMENT REQUEST
    # =========================================================================
    CreateAssignmentBffRequest:
      type: object
      description: Dati per l'assegnazione di un medico a una sede
      required:
        - doctorId
        - specialization
      properties:
        doctorId:
          type: string
          example: DOC-1
        specialization:
          type: string
          example: CARDIOLOGIA
    UpdateAssignmentBffRequest:
      type: object
      description: Dati per l'aggiornamento di un'assegnazione
      properties:
        specialization:
          type: string
        status:
          type: string

    # =========================================================================
    # APPOINTMENT REQUEST
    # =========================================================================
...